requirements.txt
# Core ML / CV
ultralytics==8.3.130
opencv-python==4.11.0.86
numpy==1.24.2
torch==2.7.0
torchvision==0.22.0
onnx==1.17.0
onnxruntime==1.22.0

# Face recognition
face-recognition==1.3.0
face-recognition-models==0.3.0
dlib==19.24.8

# Audio / Speech
sounddevice==0.5.1
vosk==0.3.45
PyAudio==0.2.13
gTTS==2.5.4   # optional, if you want TTS alternative to espeak

# GPIO / Pi
RPi.GPIO==0.7.1a4
gpiozero==2.0.1
pigpio==1.78
lgpio==0.2.2.0
picamera2==0.3.27
v4l2-python3==0.3.5

# Utils
python-dotenv==0.21.0
matplotlib==3.10.3
pandas==2.2.3
scipy==1.15.3
seaborn==0.13.2
tqdm==4.64.1

# Image I/O
Pillow==9.4.0
simplejpeg==1.8.1

# Email / GPS helpers
requests==2.28.1
geopy==2.4.1   # optional, if you add real GPS parsing

# Optional extras (remove if not needed)
reportlab==3.6.12   # for PDF generation
